requirements.txt
discord.py
jishaku
requests
bs4
pycryptodome
argon2-cffi
bcrypt